storeDetailPage 수정 => axios 분리, view 컴포넌트 생성 후 화면 구성관련 기능 정리
