+++ front/src/Page/StorePage/storeDetailPage.js
@@ -1,0 +1,16 @@
+import { useEffect, useState } from "react";
+import styled from "styled-components";
+import axios from "axios";
+import { useLocation, Link } from "react-router-dom";
+import { RiStarFill } from "react-icons/ri";
+
+import Header from "../ViewHeader";
+import Button from "../../component/Button";
+import StoreMap from "../../component/StoreMap";
+import GoodButton from "../../component/GoodButton.js";
+
+const StoreDetail = () => {
+  return;
+};
+
+export default StoreDetail;
