+++ front/src/Page/ViewMenu.js
